Basic unit test framework for TinyOS.
Provides a basic unit test framework not dissimilar to JUnit/PyUnit.  In
concept it originates from TUnit, but does not attempt to provide
regression/acceptance testing or automatic build support for multiple
configurations across multiple devices.  It just supports applications that
do assertion-based testing with a visible indicator of success/failure.  See
the README.txt in ${TOSDIR}/system/unittest.

A stub unit test in system/unittest/Stub shows the general techniques.  An
example of a fairly complex unit test is in the osian/fragpool branch, under
tos/lib/fragpool/tests.

This branch depends on osian/led for the colored pass/fail indicators, and
osian/serialprintf for the validation messages.

git-svn-id: http://tinyos-main.googlecode.com/svn/trunk@5458 285f258d-8064-b571-fbc6-9656dc998e90