Remove all built files
Files that are built do not belong in a repository. These are
appropriate for a release tarball but should not be included in a
repository. Keeping built files around causes too much churn, such as
that in #185, where the author had a different version of docutils,
which caused every built file to change.

Issue #185 should wait for this to go through and rebase to avoid the
unnecessary churn.