make new ncc (platform vs. .platform) backward compatible
We've moved .platform (and .family and .sensor) to non-dot
files so they are visible.  This requires a new version of ncc
(included in the tinyos-tools package) which isn't out yet.

In the meantime, make the new structure backward compatible with the
old ncc.   (simple symbolic links to new non-dot files).