Moved support code to tools/, now gets installed
In order to remove more environment variable usage, the tinyos python
code must be installed on the user's system instead of in the tinyos
tree. This commit moves all the python to tools so automake will install
it. Doing this requires one subtle change: "import tos" will no longer
work. It needs to be "from tinyos import tos" now.

I think I updated all of the relevant files.

One note: python3 is handled a bit weirdly, and therefore tos-serlook.py
may not be able to find tinyos.misc.hexdump. It seems that
/usr/local/lib/python3/dist-packages is not in the search path by
default. I fixed this by adding "tinyos.pth" containing the correct
directory to /usr/local/lib/python3.2/dist-packages.

The CoAP support code will now be built and installed with tinyos-tools.

I updated coap.extra to reflect the moved .c files.

Moved blip support to /tools directory

Will not compile the blip sdk code on cygwin as it only supports linux and mac.

Moved sf and serial_tun to /tools

I have no idea if anyone still uses this code. The configure script for
sf is out of date and does not work with newer automake (I updated it).

I also renamed some binaries so they can be installed. If the user is
installing TinyOS, they may as well _install_ tinyos. Why make them hunt
around for different binaries? If something shouldn't be installed, then
the question is why is it distributed with the main TinyOS source? It
should be some sort of optional extra somewhere, not part of the main
source. Hard drives are big, just give me all of the TinyOS tools.

Because there are 2 sf applications, the cpp version is named sf2.

Moved java to /tools; Removed env vars

Java is now compiled and installed with tinyos-tools

Moved imote2-ocd-program.py to /tools

No reason to have the makefile need to dig into tools in order to
install to the hardware.