Cleaned up and combined the DS2411 Code
- Changed variable and macro names. Removed anything refering to Dallas
as that doesn't make much sense anymore.
- Moved onewire to the lib folder and made a more intuitive Hpl file
in platform/chips
- Made the filenames and file functions easier to understand. It is
still 7 files to get the Eui64, but that is better than 11 and they make
more sense and are located in intuitive folders.
- Converted shimmer/2/2r/span to the global ds2411 chip
- Added an optional Ds24x1PowerControlC module that provides a
stdcontrol interface to allow platforms to start and stop the ds2411.
The version in tos/chips/ds24x1 is just a dummy version that doesn't do
anything, but a platform can override it and make the start() and stop()
commands more interesting.