BLIP printf small fix
If PRINTFUART is not defined, blip_printf.h now correctly removes all
the printf functions. printf_ieee154addr() was not being removed but
that wasn't a problem because it isn't called anywhere. But, if someone
did want to use it, their code would not compile with printf debugging
turned off.