BLIP no longer starts itself.
The way IPDispatchP was calling its own SplitControl meant that in a top
level app SplitControl.done() would be called twice: once when it was
triggered from IPDispatch and once when the top level app called
SplitControl.start().

Since BLIP provides a SplitControl interface, components that want to
use blip should be able to call start whenever they want and not have
IPDispatch automagically doing it for them.