Ieee154Address interface needs to be global
Blip is the only thing that I know of that uses the Ieee154Address
interface, but it's designed to be supported by all radio drivers. Right
now the cc2420 driver supports it, but implements it as a separate
component hanging off the side of the driver. Comments inside there
indicate that ideally the support should be built into the driver. I did
that with a different driver (not in tinyos-main) but in order to
compile it now I will need this interface (with or without blip). Hence
why it needs to be global.