Reconfigure how DHCP works with blip
While the magical #define to use DHCP or a static IP address with blip
still works, now application writers are encouraged to explicitely
declare by what mechanism they would like their application to aquire
an IPv6 address.

To use a static address based on a #define prefix (IN6_PREFIX) add:

    components StaticIPAddressC;

to your application. To use DHCP, add:

    components Dhcp6C;

to your application.

I changed the name from NoDhcpC to StaticIpAddressC because I hope to
implement SLAAC, and in that case there are two alternatives to a static
IP, and a more descriptive name is needed.