Improved pack_address() to better use hdr comp.
RFC6282 allows a node to completely elide the src address of the ipv6
header if the prefix matches the global prefix and the lower 64 bits can
be determined from other parts of the packet. This patch adds that
support by checking to see if the EUI64 that is sent in the 15.4 header
matches the lower 64 bits of the address to be packed.