BLIP: Added UDP checksum recalculation
In the case that a transmitting node sends a UDP packet and uses
stateful address compression, the UDP checksum it calculates is "wrong".
That is, it calculates it when it doesn't know its full address.
Therefore, when its address gets expanded and the UDP packet makes it to
the end host, the checksum is invalid.

This patch checks for when stateful address uncompression is used and
rechecks the checksum in that case.

This patch could be expanded to handle more headers and other checksums,
as well as a check to verify the old checksum is still correct before
creating a new one (on incorrect data at that point).