Revert the addition of BLIP_NO_RADIO
pound defines are ugly, and adding one for not using a radio in blip,
which is designed for doing wireless ip, is nonsensical. Therefore, I
removed it and added dummy files in the CoapPpp application that has the
same effect but doesn't munge up code for the sake of a single
application.