BLIP: Better handing of address modes
BLIP has relied on the magical IN6_PREFIX #define to determine if it
should use DHCP or static addresses. While this still works (to try not
to break existing code), application writers are encouraged to explicitely
declare by what mechanism they would like their application to aquire
an IPv6 address.

To use a static address based on a #define prefix (IN6_PREFIX) add:

    components StaticIPAddressC;

to your application. To use DHCP, add:

	components Dhcp6C;

to your application.

I changed the name from NoDhcpC to StaticIpAddressC because I hope to
implement SLAAC, and in that case there are two alternatives to
a static IP, and a more descriptive name is needed.

I updated the BLIP apps slightly to better select how addressing happens
while not greatly chaning it.