BLIP: Static addressing back to TOS_NODE_ID
This commits brings back the default static addressing scheme of using
the TOS_NODE_ID for the lower 64 bits. It does this by adding a new
module (StaticIPAddressTosIdC) that has the default addressing
behavior.

Now that there are three different addressing mechanisms:

 - StaticIPAddressTosIdC
 - StaticIPAddressC
 - Dhcp6C

it makes sense to have a better mechanism to make sure only one is being
used. To do this, I split out setAddress() into its own interface so that
I can use the @exactlyonce() wiring check with it.

I also removed the `#if IN6_PREFIX` check from IPStackC so there is now
no default addressing scheme. This is the only way to make this scalable
to more than two addressing schemes. Each application will have to wire
to the address assigner it would like to use.