CTP can receive message duplicates. The protocol uses the SentCache to detect duplicates: when CTP sends a message, it updates the cache record.
CTP does not update the cache when the message is received at the root
(since the root does not forward/send the message further). This
results in message duplicates at the root.

Added SentCache.insert at the root node, before it signals to the upper layer
receving of a new message. This way, message duplicates are also eliminated at
the last hop, with destination at the root.