packaging: move tinyos-tools-15 -> tinyos-tools-devel
tinyos-tools-devel builds a tinyos tools package off the development
trunk for the development trunk.  It uses the SHA from the tip
to denote the package being built coupled with the date.  Using the
date will cause a date based version to make dpkg/apt-get work
correctly with later versions of the package.

Using the SHA to identify the version clearly denotes that this is
a transient development version.

Being able to build a development verison handles the case where the
tools used to build tinyos has evolved in an incompatible fashion
and a different package needs to be built.

Note, one can always build the tools but following the instructions
in TINYOS_ROOT_DIR/tools.