packaging/tinyos-tools-14: create build script for the previous release
The tinyos-tools-14 package supports folks who insist on using the
older (previous) tinyos 2.1 release (2.1.2 etc) rather than switching
to the upcoming tinyos-2.2 release which uses the new tinyos-make v3
build system.

This is prepatory for the 2.2 release while still supporting one previous
release.

When 2.2 starts to be cut, the only tinyos-tools modifications that
need to be made are to tinyos-tools-stable.

Basically copy what is in tinyos-tools-devel with minor changes but
base it on the tarball from the tinyos-2.2 release.