tools: be much more specific about python versions
This should make sure that there is both a python2 and python3
interpreter on the system (and not just python3).

Fixes my previous commit that used the AM_PATH_PYTHON macro to check for
py2 and py3 which apparently stops at the first python interpreter it
finds, even if it doesn't satisfy the version requirement. Now we only
use that to check for py2 and explicitly look for a python3[.x] binary.