tools: specify .java files to be compiled
This fixes issue #311

tinyos.jar could be created without any class files because the
tinyos.jar target didn't have any dependencies. Because automake and
java don't play well together I think the best thing to do is to list the
.class files that need to be created explicitly and force make to build
them before creating .jar.

This is not an excellent solution, but I'm not sure anyone cares. If
there's a better way by all means change it or let me know, but I think
this will work for now.