RPL: Update the getRank() function
The RPL routing engine code calls `getRank()` to learn what rank it has
in the network. `getRank()` was implemented to accept an IP address as
an argument and return the rank of the node with that IP address. This
is fine, except when `getRank()` was called to determine the local node's
rank, it was passing the IP address the node had at boot. If the address
ever changed, `getRank()` stopped working for the local node. To avoid
this problem, a caller can just pass NULL to `getRank()` to get its own
rank.