CC2420/LPL: shorten the inter-packet interval
The inter-packet intervals in the packet trains are too long
resulting in missed transmissions as the channel probing often
happens between two packets when the channel is empty.

This fix resolves the problem by shortening CC2420_ACK_WAIT_DELAY from
256 (7.8ms) to 50 (1.5ms). Lower values break software acknowledgements,
larger may result in too long intervals.

The fix was tested both in Cooja and real TMotes increasing the
reliability of packet delivery significantly.

Note that only acknowledged unicast transmissions are affected because
broadcasts are sent much more densely (no need to wait for the ACK).