sam3: replace McuSleep/InterruptWrapper with McuSleep.irq_{pre,post}amble
The sam3 cortex code provided a Functionwrapper used as an InterruptWrapper
that allowed the interrupt handlers to get into McuSleep code that would
manipulate low power state.  This is a kludgy way to do this.

Replace by defining two new entries into McuSleep irq_pre and irq_postamble.
The irq_preamble is intended as an entrypoint into McuSleep from the Irq
handlers that then knows how to wake the CPU back up.

It is unclear what the intent of postamble is.