msp432/timer: when timer nvic interrupts are enabled enable wrap ifg
There has been a convention on the msp430 platforms to configure a timer
and immediately enable its wrap interrupt.  This is without knowledge
of what the events are driving, poor programming practice, but workable.
And it is how its been done in the past.

So we do the same thing just to get started.  This will be changed
when we implement the Msp432Timing subsystem.