panic: refactor interface, uint16_t -> parg_t
previously, motes were 16 bit machines and Panic reflected this.  With
the move to the TI msp432, we've had to address the 16 vs. 32 parameter
issue.

parg_t is introduced as panic argument type.  It is defined to be the
native underlying unsigned int type.

This commit reflects the change to using native types for panic.

(cherry picked from commit tp-master:f6c5d40)